Testplan: task end date no longer start+1 hour
The data we get from Taskjuggler now includes end dates for all
tasks. Thus, we no longer need to provide the end date for tasks as
start + 1h in the mail result. In fact that never really worked in the
first place because TJ does not accept reports for tasks that end before
the beginning of the report period.