Report last testplan
Multiple testplan instances for the same task may exist within a given
report time frame. In this case usually means someone worked on this
task. The state to report in this case is the last state, not a summary
of all tests. This patch changes the reporter to achieve this.