[TEST] fix datetime calculation for end time
This fixes rt.cpan.org #67058. Reported was a mismatch in expected and
got in the reporter unit test. This mismatch occured because we generate
the end time in "got" using DateTime but we used a fixed string in
"expected". The calucalation was thus aware of timezones while the fixed
string was not. With this patch both values are calculated (the same
way) and thus a difference should really be an error.