[FIX] Always report latest report
We used to evaluate only reports of the last week to generate a
Taskjuggler testplan report. If no test for a task was executed during
the last week for certain reasons we reported this as an error. Now we
report always the last report for a task no matter how old it is.